RENAME #17 LogicBuilder
